dev: dashboards list components, feature flag (#2380)
* dev: dashboards list components

* chore: update feature flag key

+++ web/ee/store/dashboards/workspace-dashboards.store.ts
@@ -1,6 +1,6 @@
 import set from "lodash/set";
 import unset from "lodash/unset";
-import { computed, makeObservable, observable, runInAction } from "mobx";
+import { action, computed, makeObservable, observable, runInAction } from "mobx";
 import { computedFn } from "mobx-utils";
 // plane services
 import { WorkspaceDashboardsService } from "@plane/services";
@@ -14,51 +14,72 @@ import { IWorkspaceDashboardInstance, WorkspaceDashboardInstance } from "./dashb
 
 type TLoader = "init-loader" | "mutation-loader" | undefined;
 
+type TModalPayload = (Partial<TWorkspaceDashboard> & { id: string }) | null;
+
 export interface IWorkspaceDashboardsStore {
   // observables
   loader: TLoader;
   data: Record<string, WorkspaceDashboardInstance>;
+  // modal data
+  isCreateUpdateModalOpen: boolean;
+  createUpdateModalPayload: TModalPayload;
   // helpers
   isAnyDashboardAvailable: boolean;
+  currentWorkspaceDashboardIds: string[];
   getDashboardById: (dashboardId: string) => IWorkspaceDashboardInstance | undefined;
+  toggleCreateUpdateModal: (status?: boolean) => void;
+  updateCreateUpdateModalPayload: (payload: TModalPayload) => void;
   // permissions
   canCurrentUserCreateDashboard: boolean;
   canCurrentUserDeleteDashboard: boolean;
   // actions
-  fetchDashboards: (workspaceSlug: string) => Promise<TWorkspaceDashboard[]>;
-  fetchDashboardDetails: (workspaceSlug: string, dashboardId: string) => Promise<TWorkspaceDashboard>;
-  createDashboard: (workspaceSlug: string, data: Partial<TWorkspaceDashboard>) => Promise<TWorkspaceDashboard>;
-  deleteDashboard: (workspaceSlug: string, dashboardId: string) => Promise<void>;
+  fetchDashboards: () => Promise<TWorkspaceDashboard[]>;
+  fetchDashboardDetails: (dashboardId: string) => Promise<TWorkspaceDashboard>;
+  createDashboard: (data: Partial<TWorkspaceDashboard>) => Promise<TWorkspaceDashboard>;
+  deleteDashboard: (dashboardId: string) => Promise<void>;
 }
 
 export class WorkspaceDashboardsStore implements IWorkspaceDashboardsStore {
   // observables
   loader: TLoader = "init-loader";
   data: Record<string, WorkspaceDashboardInstance> = {}; // dashboardId => DashboardInstance
+  // modal data
+  isCreateUpdateModalOpen: boolean = false;
+  createUpdateModalPayload: TModalPayload = null;
   // service
   service: WorkspaceDashboardsService;
-  // store
+  // root store
   rootStore: RootStore;
 
   constructor(store: RootStore) {
     makeObservable(this, {
       // observables
       loader: observable.ref,
+      isCreateUpdateModalOpen: observable.ref,
+      createUpdateModalPayload: observable.ref,
       data: observable,
       // computed
       isAnyDashboardAvailable: computed,
+      currentWorkspaceDashboardIds: computed,
       currentUserWorkspaceRole: computed,
       canCurrentUserCreateDashboard: computed,
       canCurrentUserDeleteDashboard: computed,
+      // actions
+      toggleCreateUpdateModal: action,
+      updateCreateUpdateModalPayload: action,
     });
-    // initialize root store
-    this.rootStore = store;
     // initialize service
     this.service = new WorkspaceDashboardsService();
+    // initialize root store
+    this.rootStore = store;
   }
 
   get isAnyDashboardAvailable() {
     return Object.values(this.data ?? {}).length > 0;
+  }
+
+  get currentWorkspaceDashboardIds() {
+    return Object.keys(this.data ?? {});
   }
 
   getDashboardById: IWorkspaceDashboardsStore["getDashboardById"] = computedFn((dashboardId) => {
@@ -82,7 +103,10 @@ export class WorkspaceDashboardsStore implements IWorkspaceDashboardsStore {
     return this.currentUserWorkspaceRole >= EUserPermissions.MEMBER;
   }
 
-  fetchDashboards: IWorkspaceDashboardsStore["fetchDashboards"] = async (workspaceSlug) => {
+  // dashboard crud
+  fetchDashboards: IWorkspaceDashboardsStore["fetchDashboards"] = async () => {
+    const workspaceSlug = this.rootStore.workspaceRoot.currentWorkspace?.slug;
+    if (!workspaceSlug) throw new Error("workspaceSlug not found");
     try {
       // update loader
       if (this.isAnyDashboardAvailable) {
@@ -110,7 +134,9 @@ export class WorkspaceDashboardsStore implements IWorkspaceDashboardsStore {
     }
   };
 
-  fetchDashboardDetails: IWorkspaceDashboardsStore["fetchDashboardDetails"] = async (workspaceSlug, dashboardId) => {
+  fetchDashboardDetails: IWorkspaceDashboardsStore["fetchDashboardDetails"] = async (dashboardId) => {
+    const workspaceSlug = this.rootStore.workspaceRoot.currentWorkspace?.slug;
+    if (!workspaceSlug) throw new Error("workspaceSlug not found");
     try {
       // make api call
       const res = await this.service.retrieve(workspaceSlug, dashboardId);
@@ -130,7 +156,9 @@ export class WorkspaceDashboardsStore implements IWorkspaceDashboardsStore {
     }
   };
 
-  createDashboard: IWorkspaceDashboardsStore["createDashboard"] = async (workspaceSlug, data) => {
+  createDashboard: IWorkspaceDashboardsStore["createDashboard"] = async (data) => {
+    const workspaceSlug = this.rootStore.workspaceRoot.currentWorkspace?.slug;
+    if (!workspaceSlug) throw new Error("workspaceSlug not found");
     try {
       // make api call
       const res = await this.service.create(workspaceSlug, data);
@@ -147,7 +175,9 @@ export class WorkspaceDashboardsStore implements IWorkspaceDashboardsStore {
     }
   };
 
-  deleteDashboard: IWorkspaceDashboardsStore["deleteDashboard"] = async (workspaceSlug, dashboardId) => {
+  deleteDashboard: IWorkspaceDashboardsStore["deleteDashboard"] = async (dashboardId) => {
+    const workspaceSlug = this.rootStore.workspaceRoot.currentWorkspace?.slug;
+    if (!workspaceSlug) throw new Error("workspaceSlug not found");
     try {
       // make api call
       await this.service.destroy(workspaceSlug, dashboardId);
@@ -159,5 +189,24 @@ export class WorkspaceDashboardsStore implements IWorkspaceDashboardsStore {
       console.error("Error in deleting workspace dashboard:", error);
       throw error;
     }
+  };
+
+  // modal helpers
+  toggleCreateUpdateModal: IWorkspaceDashboardsStore["toggleCreateUpdateModal"] = (status) => {
+    if (status === undefined) {
+      runInAction(() => {
+        this.isCreateUpdateModalOpen = !this.isCreateUpdateModalOpen;
+      });
+    } else {
+      runInAction(() => {
+        this.isCreateUpdateModalOpen = status;
+      });
+    }
+  };
+
+  updateCreateUpdateModalPayload: IWorkspaceDashboardsStore["updateCreateUpdateModalPayload"] = (payload) => {
+    runInAction(() => {
+      this.createUpdateModalPayload = payload;
+    });
   };
 }
